switch them to rest::models
