+++ include/disccord/rest/api_client.hpp
@@ -17,16 +17,16 @@
 #include <disccord/models/guild.hpp>
 #include <disccord/models/guild_embed.hpp>
 #include <disccord/models/guild_member.hpp>
-#include <disccord/models/guild_prune.hpp>
 #include <disccord/models/integration.hpp>
 #include <disccord/models/invite.hpp>
 #include <disccord/models/message.hpp>
 #include <disccord/models/role.hpp>
 #include <disccord/models/user.hpp>
-#include <disccord/models/nickname.hpp>
 #include <disccord/models/user_guild.hpp>
 #include <disccord/models/voice_region.hpp>
 
+#include <disccord/rest/models/guild_prune.hpp>
+#include <disccord/rest/models/nickname.hpp>
 #include <disccord/rest/models/add_guild_member_args.hpp>
 #include <disccord/rest/models/create_channel_invite_args.hpp>
 #include <disccord/rest/models/create_guild_channel_args.hpp>
@@ -163,7 +163,7 @@ namespace disccord
 
                     pplx::task<void> modify_guild_member(uint64_t guild_id, uint64_t user_id, disccord::rest::models::modify_guild_member_args args, const pplx::cancellation_token& token = pplx::cancellation_token::none());
 
-                    pplx::task<disccord::models::nickname> modify_current_nick(uint64_t guild_id, std::string nick, const pplx::cancellation_token& token = pplx::cancellation_token::none());
+                    pplx::task<disccord::rest::models::nickname> modify_current_nick(uint64_t guild_id, std::string nick, const pplx::cancellation_token& token = pplx::cancellation_token::none());
 
                     pplx::task<void> add_guild_member_role(uint64_t guild_id, uint64_t user_id, uint64_t role_id, const pplx::cancellation_token& token = pplx::cancellation_token::none());
 
@@ -187,9 +187,9 @@ namespace disccord
 
                     pplx::task<void> delete_guild_role(uint64_t guild_id, uint64_t role_id, const pplx::cancellation_token& token = pplx::cancellation_token::none());
 
-                    pplx::task<disccord::models::guild_prune> get_guild_prune_count(uint64_t guild_id, uint32_t days = 1, const pplx::cancellation_token& token = pplx::cancellation_token::none());
+                    pplx::task<disccord::rest::models::guild_prune> get_guild_prune_count(uint64_t guild_id, uint32_t days = 1, const pplx::cancellation_token& token = pplx::cancellation_token::none());
 
-                    pplx::task<disccord::models::guild_prune> begin_guild_prune(uint64_t guild_id, uint32_t days = 1, const pplx::cancellation_token& token = pplx::cancellation_token::none());
+                    pplx::task<disccord::rest::models::guild_prune> begin_guild_prune(uint64_t guild_id, uint32_t days = 1, const pplx::cancellation_token& token = pplx::cancellation_token::none());
 
                     pplx::task<std::vector<disccord::models::voice_region>> get_guild_voice_regions(uint64_t guild_id, const pplx::cancellation_token& token = pplx::cancellation_token::none());
