+++ lib/CMakeLists.txt
@@ -21,7 +21,7 @@ models/integration_account.cpp models/integration.cpp models/channel.cpp
 models/overwrite.cpp models/connection.cpp models/game.cpp models/presence.cpp
 models/guild_embed.cpp models/guild_member.cpp models/read_state.cpp models/relationship.cpp
 models/ban.cpp models/application.cpp models/voice_region.cpp models/voice_state.cpp
-models/nickname.cpp models/guild_prune.cpp rest/models/create_dm_channel_args.cpp
+rest/models/nickname.cpp rest/models/guild_prune.cpp rest/models/create_dm_channel_args.cpp
 rest/models/create_group_dm_args.cpp rest/models/create_message_args.cpp
 rest/models/create_guild_ban_args.cpp rest/models/modify_current_nick_args.cpp
 rest/models/modify_guild_args.cpp rest/models/modify_guild_integration_args.cpp
