+++ lib/rest/api_client.cpp
@@ -437,11 +437,11 @@ namespace disccord
                 return request_json(route, args, token);
             }
 
-            pplx::task<disccord::models::nickname> rest_api_client::modify_current_nick(uint64_t guild_id, std::string nick, const pplx::cancellation_token& token)
+            pplx::task<disccord::rest::models::nickname> rest_api_client::modify_current_nick(uint64_t guild_id, std::string nick, const pplx::cancellation_token& token)
             {
                 disccord::rest::models::modify_current_nick_args args{nick};
                 auto route = get_route("PATCH", "/guilds/{guild.id}/members/@me/nick", std::to_string(guild_id));
-                return request_json<disccord::models::nickname>(route, args, token);
+                return request_json<disccord::rest::models::nickname>(route, args, token);
             }
 
             pplx::task<void> rest_api_client::add_guild_member_role(uint64_t guild_id, uint64_t user_id, uint64_t role_id, const pplx::cancellation_token& token)
@@ -511,16 +511,16 @@ namespace disccord
                 return request(route, token);
             }
 
-            pplx::task<disccord::models::guild_prune> rest_api_client::get_guild_prune_count(uint64_t guild_id, uint32_t days, const pplx::cancellation_token& token)
+            pplx::task<disccord::rest::models::guild_prune> rest_api_client::get_guild_prune_count(uint64_t guild_id, uint32_t days, const pplx::cancellation_token& token)
             {
                 auto route = get_route("GET", "/guilds/{guild.id}/prune?days={days}", std::to_string(guild_id), std::to_string(days));
-                return request_json<disccord::models::guild_prune>(route, token);
+                return request_json<disccord::rest::models::guild_prune>(route, token);
             }
 
-            pplx::task<disccord::models::guild_prune> rest_api_client::begin_guild_prune(uint64_t guild_id, uint32_t days, const pplx::cancellation_token& token)
+            pplx::task<disccord::rest::models::guild_prune> rest_api_client::begin_guild_prune(uint64_t guild_id, uint32_t days, const pplx::cancellation_token& token)
             {
                 auto route = get_route("POST", "/guilds/{guild.id}/prune?days={days}", std::to_string(guild_id), std::to_string(days));
-                return request_json<disccord::models::guild_prune>(route, token);
+                return request_json<disccord::rest::models::guild_prune>(route, token);
             }
 
             pplx::task<std::vector<disccord::models::voice_region>> rest_api_client::get_guild_voice_regions(uint64_t guild_id, const pplx::cancellation_token& token)
